rmed steps.rs, main1.rs, updates to post.rs, fixed warnings
