+++ src/calculations.rs
@@ -7,8 +7,8 @@
 // extern crate rand;
 use crate::lib_fns;
 use rand::Rng;
-use std::io;
 use std::fmt::Write as OtherWrite;
+use std::io;
 use std::io::prelude::*;
 use std::io::Write;
 use std::process::exit; // need flush() method.
@@ -32,7 +32,7 @@ pub enum SegName {
 }
 
 // array of segment names for printing and comparison purposes.
-pub static SEG_NAMES: [&'static str; 3] = ["Code", "Heap", "Stack"];
+pub static SEG_NAMES: [&str; 3] = ["Code", "Heap", "Stack"];
 
 // shows the `student` how to solve the stack to va problem
 pub fn show_solution_stack_va(
@@ -42,39 +42,58 @@ pub fn show_solution_stack_va(
     power_of2: u32,
     percent: f32,
     aformate: i8,
-) ->String{
+) -> String {
     let mut to_print = String::new();
-    writeln(&mut to_print,"--Note that the stack base and size are parameters in Physical Memory --not Virtual Memory.");
-    writeln!(&mut to_print,"The problem prompts a translation from PA to VA --the address corresponding to the specified portion through the stack as a virtual address");
-    writeln!(&mut to_print,"Therefore, calculating the percent of the stack size and subtracting it from the stack base yields an incorrect translation.\n");
-    writeln!(&mut to_print,"Drawing a picture helps to understand this non-intuitive math.");
+    writeln!(&mut to_print,"--Note that the stack base and size are parameters in Physical Memory --not Virtual Memory.").unwrap();
+    writeln!(&mut to_print,"The problem prompts a translation from PA to VA --the address corresponding to the specified portion through the stack as a virtual address").unwrap();
+    writeln!(&mut to_print,"Therefore, calculating the percent of the stack size and subtracting it from the stack base yields an incorrect translation.\n").unwrap();
+    writeln!(
+        &mut to_print,
+        "Drawing a picture helps to understand this non-intuitive math."
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
         "\nStep 1: Calculate {}% into the stack _size_ in K (as a multiple of 1024)",
         percent
-    );
-    writeln!(&mut to_print,
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
         "{}% * {}K = ({} * {})K = {}K\n",
         percent,
         seg.size,
         percent / 100.0,
         seg.size,
         seg.size * (percent / 100.0)
-    );
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,"Step 2: Calculate the maximum segment size in relation to the stack's base");
-    writeln!(&mut to_print,"Remember --the stack grows downwards from its base.\n");
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
+        "Step 2: Calculate the maximum segment size in relation to the stack's base"
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "Remember --the stack grows downwards from its base.\n"
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
         "MSS = 2^(number of bits in the offset) = 2^{} = {} bytes",
         power_of2 - 2,
         2u32.pow(power_of2 - 2)
-    );
+    )
+    .unwrap();
     writeln!(&mut to_print,"=> If the stack segment were to use up its MSS, it would grow downwards to {}K - {}K = {}K\n",
-            seg.base, 2u32.pow(power_of2 -2) / 1024, seg.base - 2u32.pow(power_of2 -2) / 1024);
+            seg.base, 2u32.pow(power_of2 -2) / 1024, seg.base - 2u32.pow(power_of2 -2) / 1024).unwrap();
 
-    writeln!(&mut to_print,"Step 3: Calculate the offset");
-    writeln!(&mut to_print,"Subtract the address of the maximum segment size from the stack base minus the calculated portion of the stack size in K:");
-    writeln!(&mut to_print,
+    writeln!(&mut to_print, "Step 3: Calculate the offset").unwrap();
+    writeln!(&mut to_print,"Subtract the address of the maximum segment size from the stack base minus the calculated portion of the stack size in K:").unwrap();
+    writeln!(
+        &mut to_print,
         "({}K - {}K) - {}K = {}K - {}K = {}K",
         seg.base,
         seg.size * (percent / 100.0),
@@ -82,46 +101,68 @@ pub fn show_solution_stack_va(
         seg.base as f32 - seg.size * (percent / 100.0),
         seg.base - 2u32.pow(power_of2 - 2) / 1024,
         offset
-    );
-    writeln!(&mut to_print,"This is the _offset_ portion of the binary number to be constructed for the virtual address.\n");
+    )
+    .unwrap();
+    writeln!(&mut to_print,"This is the _offset_ portion of the binary number to be constructed for the virtual address.\n").unwrap();
 
-    writeln!(&mut to_print,"Step 4: Append the appropriate segment selector bits to the offset");
-    writeln!(&mut to_print,"The SS = the top two bits of the virtual address. If virtual addresses use up 17 bits, then the SS goes in the place");
-    writeln!(&mut to_print,"of the 15th and 16th bits, counting from zero.\n");
-    writeln!(&mut to_print,"Since this translation is to a stack VA, this means the segment selector bits for that address must be 11 already.");
-    writeln!(&mut to_print,"=> SS = Stack = 11");
-    writeln!(&mut to_print,"Append these two SS bits on to the highest portion of the offset binary:");
+    writeln!(
+        &mut to_print,
+        "Step 4: Append the appropriate segment selector bits to the offset"
+    )
+    .unwrap();
+    writeln!(&mut to_print,"The SS = the top two bits of the virtual address. If virtual addresses use up 17 bits, then the SS goes in the place").unwrap();
+    writeln!(
+        &mut to_print,
+        "of the 15th and 16th bits, counting from zero.\n"
+    )
+    .unwrap();
+    writeln!(&mut to_print,"Since this translation is to a stack VA, this means the segment selector bits for that address must be 11 already.").unwrap();
+    writeln!(&mut to_print, "=> SS = Stack = 11").unwrap();
+    writeln!(
+        &mut to_print,
+        "Append these two SS bits on to the highest portion of the offset binary:"
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,"  {:b}\t--SS bits", 3 << (power_of2 - 2));
-    write!(&mut to_print,"+ ");
+    writeln!(&mut to_print, "  {:b}\t--SS bits", 3 << (power_of2 - 2)).unwrap();
+    write!(&mut to_print, "+ ").unwrap();
     lib_fns::print_leading_zeros(offset * 1024, power_of2);
-    writeln!(&mut to_print,"{:b}\t--OFFSET", offset * 1024);
+    writeln!(&mut to_print, "{:b}\t--OFFSET", offset * 1024).unwrap();
+
     io::stdout().flush().unwrap(); // ensure our output is flushed entirely. print! doesnt print a line.
-    print!("--");
+
+    write!(&mut to_print, "--").unwrap();
     lib_fns::print_hyphens(3 << (power_of2 - 2));
     io::stdout().flush().unwrap();
-    writeln!(&mut to_print,"  {:b}", offset * 1024 + (3 << (power_of2 - 2)));
+    writeln!(
+        &mut to_print,
+        "  {:b}",
+        offset * 1024 + (3 << (power_of2 - 2))
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
         "{:horiz$}VA in bytes",
         " ",
         horiz = (lib_fns::num_bits_reqd(3 << (power_of2 - 2)) / 3) as usize
-    );
+    )
+    .unwrap();
 
     match aformate {
         16 => {
-            writeln!(&mut to_print,"=> VA = {:#X} bytes", va);
+            writeln!(&mut to_print, "=> VA = {:#X} bytes", va).unwrap();
         }
         2 => {}
         10 => {
-            writeln!(&mut to_print,"=> VA = {} bytes", va);
+            writeln!(&mut to_print, "=> VA = {} bytes", va).unwrap();
         }
         _ => {
             error();
         }
     }
-    writeln!(&mut to_print,"Check out youtube for shortcuts on converting to and from binary, decimal, and hexadecimal by hand.\n");
-to_print
+    writeln!(&mut to_print,"Check out youtube for shortcuts on converting to and from binary, decimal, and hexadecimal by hand.\n").unwrap();
+    to_print
 }
 
 // shows the `student` the steps to solving the VA to PA problem
@@ -133,149 +174,263 @@ pub fn show_solution_va_to_pa_hex(
     pa: u32,
     power_of2: u32,
     qaformate: (i8, i8),
-) ->String {
-    let mut to_print = String:new();
+) -> String {
+    let mut to_print = String::new();
     match qaformate.0 {
         16 => {
-            writeln!(&mut to_print,"\nStep 1: Convert virtual address {:#X} to binary", va);
-            write!(&mut to_print,"{:#X} = ", va);
+            writeln!(
+                &mut to_print,
+                "\nStep 1: Convert virtual address {:#X} to binary",
+                va
+            )
+            .unwrap();
+            write!(&mut to_print, "{:#X} = ", va).unwrap();
         }
         2 => {
-            writeln!(&mut to_print,"\nStep 1: Convert virtual address {:b} to binary", va);
-            print!("{:b} = ", va);
+            writeln!(
+                &mut to_print,
+                "\nStep 1: Convert virtual address {:b} to binary",
+                va
+            )
+            .unwrap();
+            write!(&mut to_print, "{:b} = ", va).unwrap();
         }
         10 => {
-            writeln!(&mut to_print,"\nStep 1: Convert virtual address {} to binary", va);
-            write!(&mut to_print,"{} = ", va);
+            writeln!(
+                &mut to_print,
+                "\nStep 1: Convert virtual address {} to binary",
+                va
+            )
+            .unwrap();
+            write!(&mut to_print, "{} = ", va).unwrap();
         }
         _ => {
             error();
         }
-
     }
     // lib_fns::print_leading_zeros(va, power_of2);
     // print!("{:b} = ", va);
     lib_fns::print_readable_binary(va, power_of2); // I think this function is cool, visit the file to check it out.
-    writeln!(&mut to_print,"");
+    writeln!(&mut to_print).unwrap();
     io::stdout().flush().unwrap(); // ensure our output is flushed entirely, as we are not using the _println_ macro here
 
-    writeln!(&mut to_print,"\nStep 2: Note the Virtual Address Space size (abbrv. VAS --measured in bits) and separate the Segment Selector (SS) from the Offset portion of the binary.");
-    writeln!(&mut to_print,"Remember --if the amount of bits in the Virtual Address Space differs from the amount of bits in the binary calculated, we must either");
-    writeln!(&mut to_print,"\n\ta\u{29} Pad the calculated binary number with zeros until the length of the binary equals the amount of bits in the VAS.");
-    writeln!(&mut to_print,"\n\tb\u{29} Trim the top bits of the calculated binary until the length of the binary equals the amount of bits in the VAS.\n");
-    writeln!(&mut to_print,
+    writeln!(&mut to_print,"\nStep 2: Note the Virtual Address Space size (abbrv. VAS --measured in bits) and separate the Segment Selector (SS) from the Offset portion of the binary.").unwrap();
+    writeln!(&mut to_print,"Remember --if the amount of bits in the Virtual Address Space differs from the amount of bits in the binary calculated, we must either").unwrap();
+    writeln!(&mut to_print,"\n\ta\u{29} Pad the calculated binary number with zeros until the length of the binary equals the amount of bits in the VAS.").unwrap();
+    writeln!(&mut to_print,"\n\tb\u{29} Trim the top bits of the calculated binary until the length of the binary equals the amount of bits in the VAS.\n").unwrap();
+    writeln!(
+        &mut to_print,
         "In this case, the Virtual Address Space size in bits is {}.",
         power_of2
-    );
-    writeln!(&mut to_print,
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
         "So, only the first {} bits of the calculated binary are considered.\n",
         power_of2
-    );
-    writeln!(&mut to_print,"The SS is always either 00, 01, or 11 => SS \u{3F5} \u{7B}{0, 1, 3\u{7D}}");
-    writeln!(&mut to_print,"\nDiscard the segment selector bits from the offset calculation.\n");
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "The SS is always either 00, 01, or 11 => SS \u{3F5} \u{7B}{0, 1, 3\u{7D}}"
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "\nDiscard the segment selector bits from the offset calculation.\n"
+    )
+    .unwrap();
     if ss != 3 {
-        writeln!(&mut to_print,"0{} {:b}", ss, offset);
+        writeln!(&mut to_print, "0{} {:b}", ss, offset).unwrap();
     } else {
-        writeln!(&mut to_print,"{:b} {:b}", ss, offset);
+        writeln!(&mut to_print, "{:b} {:b}", ss, offset).unwrap();
     }
-    write!(&mut to_print,"-- ");
+    write!(&mut to_print, "-- ").unwrap();
     lib_fns::print_hyphens(offset);
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
         "SS {:x_axis$}OFFSET",
         " ",
         x_axis = ((lib_fns::num_bits_reqd(offset) / 2) / 2) as usize
-    );
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,"\nStep 3: Note the value of the Segment Selector and Offset bits:");
-    writeln!(&mut to_print,"00 ===> Code");
-    writeln!(&mut to_print,"01 ===> Heap");
-    writeln!(&mut to_print,"11 ===> Stack\n");
-    writeln!(&mut to_print,"Offset = {:b} = {} bytes\n", offset, offset);
+    writeln!(
+        &mut to_print,
+        "\nStep 3: Note the value of the Segment Selector and Offset bits:"
+    )
+    .unwrap();
+    writeln!(&mut to_print, "00 ===> Code").unwrap();
+    writeln!(&mut to_print, "01 ===> Heap").unwrap();
+    writeln!(&mut to_print, "11 ===> Stack\n").unwrap();
+    writeln!(&mut to_print, "Offset = {:b} = {} bytes\n", offset, offset).unwrap();
 
-    writeln!(&mut to_print,"Step 4: Note: PA = (-1)*(GN)*(MSS) + base + offset\n");
-    writeln!(&mut to_print,"GN = `grows negative`. If SS = 11 => SS = Stack => GN = 1. Otherwise, GN = 0.");
-    writeln!(&mut to_print,"MSS = `maximum segment size`. MSS = 2^(number of bits in the offset)");
-    writeln!(&mut to_print,"Base = the base of the segment, measured in bytes. Value provided in table.");
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
+        "Step 4: Note: PA = (-1)*(GN)*(MSS) + base + offset\n"
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "GN = `grows negative`. If SS = 11 => SS = Stack => GN = 1. Otherwise, GN = 0."
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "MSS = `maximum segment size`. MSS = 2^(number of bits in the offset)"
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
+        "Base = the base of the segment, measured in bytes. Value provided in table."
+    )
+    .unwrap();
+    writeln!(
+        &mut to_print,
         "The offset has already been calculated: offset = {} bytes (base 10)",
         offset
-    );
+    )
+    .unwrap();
 
-    writeln!(&mut to_print,
+    writeln!(
+        &mut to_print,
         "There are {} bits in the offset, so the MSS is 2^{} = {} bytes.",
         power_of2 - 2,
         power_of2 - 2,
         2u32.pow(power_of2 - 2)
-    );
+    )
+    .unwrap();
     if ss == 0 {
-        writeln!(&mut to_print,"The SS = 00 (base 2) = 0 (base 10) => SS = Code Segment => GN = 0");
-        writeln!(&mut to_print,"PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n");
-        writeln!(&mut to_print,
+        writeln!(
+            &mut to_print,
+            "The SS = 00 (base 2) = 0 (base 10) => SS = Code Segment => GN = 0"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(0)(2^{}) + ({}K) + {} bytes",
             power_of2 - 2,
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(0)({}) + ({} * 1024) + {} bytes",
             2u32.pow(power_of2 - 2),
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,"=> PA = 0 + ({}) + {} bytes", seg.base * 1024, offset);
-        writeln!(&mut to_print,"=> PA = {} bytes", (seg.base * 1024) + offset);
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "=> PA = 0 + ({}) + {} bytes",
+            seg.base * 1024,
+            offset
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "=> PA = {} bytes",
+            (seg.base * 1024) + offset
+        )
+        .unwrap();
         if pa != (seg.base * 1024) + offset {
-            writeln!(&mut to_print,"Error. Conflicting calculations of pa");
+            writeln!(&mut to_print, "Error. Conflicting calculations of pa").unwrap();
             exit(-1);
         }
     } else if ss == 1 {
-        writeln!(&mut to_print,"The SS = 01 (base 2) = 1 (base 10) => SS = Heap Segment => GN = 0");
-        writeln!(&mut to_print,"PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n");
-        writeln!(&mut to_print,
+        writeln!(
+            &mut to_print,
+            "The SS = 01 (base 2) = 1 (base 10) => SS = Heap Segment => GN = 0"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(0)(2^{}) + ({}K) + {} bytes",
             power_of2 - 2,
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(0)({}) + ({} * 1024) + {} bytes",
             2u32.pow(power_of2 - 2),
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,"=> PA = 0 + ({}) + {} bytes", seg.base * 1024, offset);
-        writeln!(&mut to_print,"=> PA = {} bytes", (seg.base * 1024) + offset);
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "=> PA = 0 + ({}) + {} bytes",
+            seg.base * 1024,
+            offset
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "=> PA = {} bytes",
+            (seg.base * 1024) + offset
+        )
+        .unwrap();
         if pa != (seg.base * 1024) + offset {
-            writeln!(&mut to_print,"Error. Conflicting calculations of pa");
+            writeln!(&mut to_print, "Error. Conflicting calculations of pa").unwrap();
             exit(-1);
         }
     } else if ss == 3 {
-        writeln!(&mut to_print,"The SS = 11 (base 2) = 3 (base 10) => SS = Stack Segment => GN = 1");
-        writeln!(&mut to_print,"PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n");
-        writeln!(&mut to_print,
+        writeln!(
+            &mut to_print,
+            "The SS = 11 (base 2) = 3 (base 10) => SS = Stack Segment => GN = 1"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
+            "PA (in bytes) = (-1)*(GN)*(MSS) + base + offset\n"
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(1)(2^{}) + ({}K) + {} bytes",
             power_of2 - 2,
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-1)(1)({}) + ({} * 1024) + {} bytes",
             2u32.pow(power_of2 - 2),
             seg.base,
             offset
-        );
-        writeln!(&mut to_print,
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = (-{}) + ({}) + {} bytes",
             2u32.pow(power_of2 - 2),
             seg.base * 1024,
             offset
-        );
-        writeln!(&mut to_print,
+        )
+        .unwrap();
+        writeln!(
+            &mut to_print,
             "=> PA = {} bytes",
             ((seg.base * 1024) + offset) - 2u32.pow(power_of2 - 2)
-        );
+        )
+        .unwrap();
         if pa != (seg.base * 1024) + offset - 2u32.pow(power_of2 - 2) {
-            writeln!(&mut to_print,"Error. Conflicting calculations of pa");
+            writeln!(&mut to_print, "Error. Conflicting calculations of pa").unwrap();
             exit(-1);
         }
     } else {
@@ -284,18 +439,18 @@ pub fn show_solution_va_to_pa_hex(
 
     match qaformate.1 {
         16 => {
-            writeln!(&mut to_print,"=> PA = {:#X} bytes", pa);
+            writeln!(&mut to_print, "=> PA = {:#X} bytes", pa).unwrap();
         }
         2 => {
-            writeln!(&mut to_print,"=> PA = {:b} bytes", pa);
+            writeln!(&mut to_print, "=> PA = {:b} bytes", pa).unwrap();
         }
         10 => {}
         _ => {
             error();
         }
     }
-    writeln!(&mut to_print,"Check out youtube for shortcuts on converting to and from binary, decimal, and hexadecimal by hand.\n");
-to_print
+    writeln!(&mut to_print,"Check out youtube for shortcuts on converting to and from binary, decimal, and hexadecimal by hand.\n").unwrap();
+    to_print
 }
 
 // this function calculates the answer to the stack percentage problem.
@@ -326,7 +481,10 @@ pub fn test_stack_percentage_calculation_v1() {
         size: 2.0,
         grows_negative: 1,
     };
-    assert_eq!(114688, calculate_answer_stack_percentage(stak_seg, 0.0, 16384, 17).0);
+    assert_eq!(
+        114688,
+        calculate_answer_stack_percentage(stak_seg, 0.0, 16384, 17).0
+    );
 }
 
 #[test]
@@ -337,7 +495,10 @@ pub fn test_stack_percentage_calculation_v2() {
         size: 0.0,
         grows_negative: 1,
     };
-    assert_eq!(196608, calculate_answer_stack_percentage(stak_seg, 0.0, 16384, 18).0);
+    assert_eq!(
+        196608,
+        calculate_answer_stack_percentage(stak_seg, 0.0, 16384, 18).0
+    );
 }
 
 // this function calculates the answer to the given problem and returns the result
@@ -345,9 +506,10 @@ pub fn test_stack_percentage_calculation_v2() {
 // GN = grows negative. offset = the value of the virtual address shifted left 2 bits.
 // (the value of the va neglecting the final two bits)
 #[allow(clippy::neg_multiply)] // --Massey suggestion to get rid of clippy warning.
-// using (1 * (!1)) does not fix the issue, it results in -2 instead of -1
+                               // using (1 * (!1)) does not fix the issue, it results in -2 instead of -1
 pub fn calculate_answer(seg: Segment, mss: u32, offset: u32) -> u32 {
-    ((-1) * (seg.grows_negative as i32) * (mss as i32) + (seg.base as i32 * 1024) + offset as i32) as u32 // in bytes
+    ((-1) * (seg.grows_negative as i32) * (mss as i32) + (seg.base as i32 * 1024) + offset as i32)
+        as u32 // in bytes
 }
 
 #[test]
@@ -364,15 +526,16 @@ pub fn test_va_pa_calculation_v1() {
 // compares actual answer to user answer after printing the question
 pub fn compare_answer(aformat: i8, pa: u32) {
     let mut input = String::new();
+    let mut to_print = String::new();
     match aformat {
         16 => {
-            writeln!(&mut to_print,"Type your answer in hexadecimal format with or without the `0x` then press enter and ctrl+d");
+            writeln!(&mut to_print,"Type your answer in hexadecimal format with or without the `0x` then press enter and ctrl+d").unwrap();
             // the read_to_string writes the input data to its argument, not the return value.
             input = match io::stdin().read_to_string(&mut input) {
                 Ok(_usize_bytes) => input.to_string().trim().to_string(),
                 Err(_) => "".to_string(),
             };
-            writeln!(&mut to_print,"");
+            writeln!(&mut to_print).unwrap();
             input = input.replace("x", ""); // replace all the characters that could possibly be taken as hex prefixes (like 0, x) with empty string
             input = input.replace("X", "");
             if lib_fns::are_all_numeric(&input, 16) {
@@ -380,30 +543,44 @@ pub fn compare_answer(aformat: i8, pa: u32) {
                     // use my library function to convert input to base 10 (so we can measure it!)
                     Some(k) => {
                         if k as u32 == pa {
-                            writeln!(&mut to_print,"Good.");
+                            writeln!(&mut to_print, "Good.").unwrap();
                         } else {
-                            writeln!(&mut to_print,"INCORRECT.\n");
-                            writeln!(&mut to_print,"your answer: {:#X} bytes\nactual: {:#X} bytes", k, pa);
+                            writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                            writeln!(
+                                &mut to_print,
+                                "your answer: {:#X} bytes\nactual: {:#X} bytes",
+                                k, pa
+                            )
+                            .unwrap();
                         }
                     }
                     None => {
-                        writeln!(&mut to_print,"INCORRECT.\n");
-                        writeln!(&mut to_print,"your answer: {} bytes\nactual: {:#X} bytes", input, pa);
+                        writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                        writeln!(
+                            &mut to_print,
+                            "your answer: {} bytes\nactual: {:#X} bytes",
+                            input, pa
+                        )
+                        .unwrap();
                     }
                 }
             } else {
-                writeln!(&mut to_print,"INCORRECT.\n");
-                writeln!(&mut to_print,"your answer: {} bytes\nactual: {:#X} bytes", input, pa);
+                writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                writeln!(
+                    &mut to_print,
+                    "your answer: {} bytes\nactual: {:#X} bytes",
+                    input, pa
+                )
+                .unwrap();
             }
-            return;
         }
         2 => {
-            writeln!(&mut to_print,"Type your answer in binary format with or without leading zeros then press enter and ctrl+d");
+            writeln!(&mut to_print,"Type your answer in binary format with or without leading zeros then press enter and ctrl+d").unwrap();
             input = match io::stdin().read_to_string(&mut input) {
                 Ok(_usize_bytes) => input.to_string().trim().to_string(),
                 Err(_) => "".to_string(),
             };
-            writeln!(&mut to_print,"");
+            writeln!(&mut to_print).unwrap();
             input = input.replace("x", "");
             input = input.replace("X", ""); // replace all the characters that could possibly be taken as hex prefixes (like 0, x) with empty string
             if lib_fns::are_all_numeric(&input, 2) {
@@ -411,59 +588,88 @@ pub fn compare_answer(aformat: i8, pa: u32) {
                 match lib_fns::bn_to_b10(&input.trim().to_string(), 2) {
                     Some(k) => {
                         if k as u32 == pa {
-                            writeln!(&mut to_print,"Good.");
+                            writeln!(&mut to_print, "Good.").unwrap();
                         } else {
-                            writeln!(&mut to_print,"INCORRECT.\n");
-                            writeln!(&mut to_print,
+                            writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                            writeln!(
+                                &mut to_print,
                                 "your answer: {:b} bytes bytes\nactual: {:b} bytes bytes",
                                 k, pa
-                            );
+                            )
+                            .unwrap();
                         }
                     }
                     None => {
-                        writeln!(&mut to_print,"INCORRECT.\n");
-                        writeln!(&mut to_print,"your answer: {} bytes\nactual: {:b} bytes", input, pa);
+                        writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                        writeln!(
+                            &mut to_print,
+                            "your answer: {} bytes\nactual: {:b} bytes",
+                            input, pa
+                        )
+                        .unwrap();
                     }
                 }
             } else {
-                writeln!(&mut to_print,"INCORRECT.\n");
-                writeln!(&mut to_print,"your answer: {} bytes\nactual: {:b} bytes", input, pa);
+                writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                writeln!(
+                    &mut to_print,
+                    "your answer: {} bytes\nactual: {:b} bytes",
+                    input, pa
+                )
+                .unwrap();
             }
-            return;
         }
         10 => {
-            writeln!(&mut to_print,"Type your answer in decimal format (base 10, no decimal points) then press enter and ctrl+d");
+            writeln!(&mut to_print,"Type your answer in decimal format (base 10, no decimal points) then press enter and ctrl+d").unwrap();
             // the read_to_string writes the input data to its argument, not the return value.
             input = match io::stdin().read_to_string(&mut input) {
                 Ok(_usize_bytes) => input.to_string().trim().to_string(),
                 Err(_) => "".to_string(),
             };
-            writeln!(&mut to_print,"");
+            writeln!(&mut to_print).unwrap();
             input = input.replace("x", "");
             input = input.replace("X", "");
             if lib_fns::are_all_numeric(&input, 10) {
                 match lib_fns::bn_to_b10(&input.trim().to_string(), 10) {
                     Some(k) => {
                         if k as u32 == pa {
-                            writeln!(&mut to_print,"Good.");
+                            writeln!(&mut to_print, "Good.").unwrap();
                         } else {
-                            writeln!(&mut to_print,"INCORRECT.\n");
-                            writeln!(&mut to_print,"your answer: {} bytes\nactual: {} bytes", k, pa);
+                            writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                            writeln!(
+                                &mut to_print,
+                                "your answer: {} bytes\nactual: {} bytes",
+                                k, pa
+                            )
+                            .unwrap();
                         }
                     }
                     None => {
-                        writeln!(&mut to_print,"INCORRECT.\n");
-                        writeln!(&mut to_print,"your answer: {} bytes\nactual: {} bytes", input, pa);
+                        writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                        writeln!(
+                            &mut to_print,
+                            "your answer: {} bytes\nactual: {} bytes",
+                            input, pa
+                        )
+                        .unwrap();
                     }
                 }
             } else {
-                writeln!(&mut to_print,"INCORRECT.\n");
-                writeln!(&mut to_print,"your answer: {} bytes\nactual: {} bytes", input, pa);
+                writeln!(&mut to_print, "INCORRECT.\n").unwrap();
+                writeln!(
+                    &mut to_print,
+                    "your answer: {} bytes\nactual: {} bytes",
+                    input, pa
+                )
+                .unwrap();
             }
-            return;
         }
         _ => {
-            writeln!(&mut to_print,"Error. Unexpected format specifier. Fatal error. Terminating program");
+            writeln!(
+                &mut to_print,
+                "Error. Unexpected format specifier. Fatal error. Terminating program"
+            )
+            .unwrap();
             exit(-1)
         }
     }
@@ -529,7 +735,7 @@ pub fn valid_va(num: u32, fresh_ss: u32, segments: Vec<Segment>) -> bool {
 }
 
 pub fn error() {
-    writeln!(&mut to_print,"Encountered fatal error. Exiting");
+    println!("Encountered fatal error. Exiting");
     exit(-1);
 }
